Revert "Refactor tests (#9)" (#10)
This reverts commit 07a8120a3cc6ef6ab877cc65ce05c23098337cc5.
