--- old_tests/utils.py
@@ -1,34 +1,0 @@
-from brownie import Contract, chain
-
-
-def sync_price(joint, mock_chainlink, strategist):
-    pair = Contract(joint.pair())
-    (reserve0, reserve1, a) = pair.getReserves()
-    mock_chainlink.setPrice(reserve0 / reserve1 * 1e12 * 10 ** 8, {"from": strategist})
-
-
-def print_hedge_status(joint, tokenA, tokenB):
-    callID = joint.activeCallID()
-    putID = joint.activePutID()
-    callProvider = Contract("0xb9ed94c6d594b2517c4296e24A8c517FF133fb6d")
-    putProvider = Contract("0x790e96E7452c3c2200bbCAA58a468256d482DD8b")
-    callInfo = callProvider.options(callID)
-    putInfo = putProvider.options(putID)
-    assert (joint.activeCallID() != 0) & (joint.activePutID() != 0)
-    (callPayout, putPayout) = joint.getHedgeProfit()
-    print(f"Bought two options:")
-    print(f"CALL #{callID}")
-    print(f"\tStrike {callInfo[1]/1e8}")
-    print(f"\tAmount {callInfo[2]/1e18}")
-    print(f"\tTTM {(callInfo[4]-chain.time())/3600}h")
-    costCall = (callInfo[5] + callInfo[6]) / 0.8
-    print(f"\tCost {(callInfo[5]+callInfo[6])/0.8/1e18} {tokenA.symbol()}")
-    print(f"\tPayout: {callPayout/1e18} {tokenA.symbol()}")
-    print(f"PUT #{putID}")
-    print(f"\tStrike {putInfo[1]/1e8}")
-    print(f"\tAmount {putInfo[2]/1e18}")
-    print(f"\tTTM {(putInfo[4]-chain.time())/3600}h")
-    costPut = (putInfo[5] + putInfo[6]) / 0.8
-    print(f"\tCost {costPut/1e6} {tokenB.symbol()}")
-    print(f"\tPayout: {putPayout/1e6} {tokenB.symbol()}")
-    return (costCall, costPut)
