+++ requirements-dev.txt
@@ -1,2 +1,1 @@
-black==21.9b0
 eth-brownie>=1.16.3,<2.0.0
