+++ tests/spirit/test_spirit_operation.py
