+++ tests/test_operation.py
@@ -1,114 +1,273 @@
 import brownie
-from brownie import Contract
 import pytest
-from utils import actions, checks
+from brownie import Contract, Wei
+from operator import xor
+from utils import sync_price, print_hedge_status
 
 
 def test_operation(
-    chain, accounts, token, vault, strategy, user, strategist, amount, RELATIVE_APPROX
+    chain,
+    vaultA,
+    vaultB,
+    tokenA,
+    tokenB,
+    amountA,
+    amountB,
+    providerA,
+    providerB,
+    joint,
+    strategist,
+    tokenA_whale,
+    tokenB_whale,
+    mock_chainlink,
 ):
-    # Deposit to the vault
-    user_balance_before = token.balanceOf(user)
-    actions.user_deposit(user, vault, token, amount)
+    sync_price(joint, mock_chainlink, strategist)
 
-    # harvest
-    chain.sleep(1)
-    strategy.harvest({"from": strategist})
-    assert pytest.approx(strategy.estimatedTotalAssets(), rel=RELATIVE_APPROX) == amount
+    tokenA.approve(vaultA, 2 ** 256 - 1, {"from": tokenA_whale})
+    vaultA.deposit(amountA, {"from": tokenA_whale})
 
-    # tend()
-    strategy.tend({"from": strategist})
+    tokenB.approve(vaultB, 2 ** 256 - 1, {"from": tokenB_whale})
+    vaultB.deposit(amountB, {"from": tokenB_whale})
 
-    # withdrawal
-    vault.withdraw({"from": user})
-    assert (
-        pytest.approx(token.balanceOf(user), rel=RELATIVE_APPROX) == user_balance_before
+    ppsA_start = vaultA.pricePerShare()
+    ppsB_start = vaultB.pricePerShare()
+
+    providerA.harvest({"from": strategist})
+    providerB.harvest({"from": strategist})
+    print_hedge_status(joint, tokenA, tokenB)
+    assert joint.balanceOfA() == 0
+    assert joint.balanceOfB() == 0
+    assert joint.balanceOfStake() > 0
+
+    investedA = (
+        vaultA.strategies(providerA).dict()["totalDebt"] - providerA.balanceOfWant()
+    )
+    investedB = (
+        vaultB.strategies(providerB).dict()["totalDebt"] - providerB.balanceOfWant()
     )
 
+    print(
+        f"Joint estimated assets: {joint.estimatedTotalAssetsInToken(tokenA) / 1e18} {tokenA.symbol()} and {joint.estimatedTotalAssetsInToken(tokenB) / 1e18} {tokenB.symbol()}"
+    )
 
-def test_emergency_exit(
-    chain, accounts, token, vault, strategy, user, strategist, amount, RELATIVE_APPROX
+    # Wait plz
+    chain.sleep(3600 * 24 - 30)
+    chain.mine(int(3600 / 13) * 24)
+
+    print(
+        f"Joint estimated assets: {joint.estimatedTotalAssetsInToken(tokenA) / 1e18} {tokenA.symbol()} and {joint.estimatedTotalAssetsInToken(tokenB) / 1e18} {tokenB.symbol()}"
+    )
+
+    # If there is any profit it should go to the providers
+    assert joint.pendingReward() > 0
+    # If joint doesn't reinvest, and providers do not invest want, the want
+    # will stay in the providers
+    providerA.setInvestWant(False, {"from": strategist})
+    providerB.setInvestWant(False, {"from": strategist})
+    providerA.setTakeProfit(True, {"from": strategist})
+    providerB.setTakeProfit(True, {"from": strategist})
+    providerA.harvest({"from": strategist})
+    providerB.harvest({"from": strategist})
+    assert providerA.balanceOfWant() > 0
+    assert providerB.balanceOfWant() > 0
+
+    # Harvest should be a no-op
+    providerA.harvest({"from": strategist})
+    providerB.harvest({"from": strategist})
+    chain.sleep(60 * 60 * 8)
+    chain.mine(1)
+    assert providerA.balanceOfWant() > 0
+    assert providerB.balanceOfWant() > 0
+
+    chain.sleep(60 * 60 * 8)
+    chain.mine(1)
+    # losses due to not being able to earn enough to cover hedge without trades!
+    assert vaultA.strategies(providerA).dict()["totalLoss"] > 0
+    assert vaultB.strategies(providerB).dict()["totalLoss"] > 0
+
+
+def test_operation_swap_a4b(
+    chain,
+    vaultA,
+    vaultB,
+    tokenA,
+    tokenB,
+    amountA,
+    amountB,
+    providerA,
+    providerB,
+    joint,
+    router,
+    strategist,
+    tokenA_whale,
+    tokenB_whale,
+    mock_chainlink,
 ):
-    # Deposit to the vault
-    actions.user_deposit(user, vault, token, amount)
-    chain.sleep(1)
-    strategy.harvest({"from": strategist})
-    assert pytest.approx(strategy.estimatedTotalAssets(), rel=RELATIVE_APPROX) == amount
+    sync_price(joint, mock_chainlink, strategist)
+    tokenA.approve(vaultA, 2 ** 256 - 1, {"from": tokenA_whale})
+    vaultA.deposit(amountA, {"from": tokenA_whale})
 
-    # set emergency and exit
-    strategy.setEmergencyExit()
-    chain.sleep(1)
-    strategy.harvest({"from": strategist})
-    assert strategy.estimatedTotalAssets() < amount
+    tokenB.approve(vaultB, 2 ** 256 - 1, {"from": tokenB_whale})
+    vaultB.deposit(amountB, {"from": tokenB_whale})
+
+    providerA.harvest({"from": strategist})
+    providerB.harvest({"from": strategist})
+
+    assert joint.balanceOfA() == 0
+    assert joint.balanceOfB() == 0
+    assert joint.balanceOfStake() > 0
+
+    investedA = (
+        vaultA.strategies(providerA).dict()["totalDebt"] - providerA.balanceOfWant()
+    )
+    investedB = (
+        vaultB.strategies(providerB).dict()["totalDebt"] - providerB.balanceOfWant()
+    )
+    print(
+        f"Joint estimated assets: {joint.estimatedTotalAssetsInToken(tokenA) / 1e18} {tokenA.symbol()} and {joint.estimatedTotalAssetsInToken(tokenB) / 1e18} {tokenB.symbol()}"
+    )
+
+    tokenA.approve(router, 2 ** 256 - 1, {"from": tokenA_whale})
+    router.swapExactTokensForTokens(
+        tokenA.balanceOf(tokenA_whale),
+        0,
+        [tokenA, tokenB],
+        tokenA_whale,
+        2 ** 256 - 1,
+        {"from": tokenA_whale},
+    )
+
+    print(
+        f"Joint estimated assets: {joint.estimatedTotalAssetsInToken(tokenA) / 1e18} {tokenA.symbol()} and {joint.estimatedTotalAssetsInToken(tokenB) / 1e18} {tokenB.symbol()}"
+    )
+
+    # Wait plz
+    chain.sleep(3600 * 24 - 30)
+    chain.mine(int(3600 * 24 / 13) - 30)
+
+    print(
+        f"Joint estimated assets: {joint.estimatedTotalAssetsInToken(tokenA) / 1e18} {tokenA.symbol()} and {joint.estimatedTotalAssetsInToken(tokenB) / 1e18} {tokenB.symbol()}"
+    )
+
+    # If there is any profit it should go to the providers
+    assert joint.pendingReward() > 0
+    # If joint doesn't reinvest, and providers do not invest want, the want
+    # will stay in the providers
+    providerA.setInvestWant(False, {"from": strategist})
+    providerB.setInvestWant(False, {"from": strategist})
+    providerA.setTakeProfit(True, {"from": strategist})
+    providerB.setTakeProfit(True, {"from": strategist})
+    providerA.harvest({"from": strategist})
+    providerB.harvest({"from": strategist})
+
+    assert providerA.balanceOfWant() > 0
+    assert providerB.balanceOfWant() > 0
+
+    lossA = vaultA.strategies(providerA).dict()["totalLoss"]
+    lossB = vaultB.strategies(providerB).dict()["totalLoss"]
+
+    assert lossA > 0
+    assert lossB > 0
+
+    returnA = -lossA / investedA
+    returnB = -lossB / investedB
+
+    print(
+        f"Return: {returnA*100:.5f}% {tokenA.symbol()} {returnB*100:.5f}% {tokenB.symbol()}"
+    )
+
+    assert pytest.approx(returnA, rel=50e-3) == returnB
 
 
-def test_increase_debt_ratio(
-    chain, gov, token, vault, strategy, user, strategist, amount, RELATIVE_APPROX
+def test_operation_swap_b4a(
+    chain,
+    vaultA,
+    vaultB,
+    tokenA,
+    tokenB,
+    amountA,
+    amountB,
+    providerA,
+    providerB,
+    joint,
+    router,
+    strategist,
+    tokenA_whale,
+    tokenB_whale,
+    mock_chainlink,
 ):
-    # Deposit to the vault and harvest
-    actions.user_deposit(user, vault, token, amount)
-    vault.updateStrategyDebtRatio(strategy.address, 5_000, {"from": gov})
-    chain.sleep(1)
-    strategy.harvest({"from": strategist})
-    half = int(amount / 2)
+    sync_price(joint, mock_chainlink, strategist)
 
-    assert pytest.approx(strategy.estimatedTotalAssets(), rel=RELATIVE_APPROX) == half
+    tokenA.approve(vaultA, 2 ** 256 - 1, {"from": tokenA_whale})
+    vaultA.deposit(amountA, {"from": tokenA_whale})
 
-    vault.updateStrategyDebtRatio(strategy.address, 10_000, {"from": gov})
-    chain.sleep(1)
-    strategy.harvest({"from": strategist})
-    assert pytest.approx(strategy.estimatedTotalAssets(), rel=RELATIVE_APPROX) == amount
+    tokenB.approve(vaultB, 2 ** 256 - 1, {"from": tokenB_whale})
+    vaultB.deposit(amountB, {"from": tokenB_whale})
 
+    providerA.harvest({"from": strategist})
+    providerB.harvest({"from": strategist})
 
-def test_decrease_debt_ratio(
-    chain, gov, token, vault, strategy, user, strategist, amount, RELATIVE_APPROX
-):
-    # Deposit to the vault and harvest
-    actions.user_deposit(user, vault, token, amount)
-    vault.updateStrategyDebtRatio(strategy.address, 10_000, {"from": gov})
-    chain.sleep(1)
-    strategy.harvest({"from": strategist})
+    assert joint.balanceOfA() == 0
+    assert joint.balanceOfB() == 0
+    assert joint.balanceOfStake() > 0
 
-    assert pytest.approx(strategy.estimatedTotalAssets(), rel=RELATIVE_APPROX) == amount
+    investedA = (
+        vaultA.strategies(providerA).dict()["totalDebt"] - providerA.balanceOfWant()
+    )
+    investedB = (
+        vaultB.strategies(providerB).dict()["totalDebt"] - providerB.balanceOfWant()
+    )
+    print(
+        f"Joint estimated assets: {joint.estimatedTotalAssetsInToken(tokenA) / 1e18} {tokenA.symbol()} and {joint.estimatedTotalAssetsInToken(tokenB) / 1e18} {tokenB.symbol()}"
+    )
 
-    vault.updateStrategyDebtRatio(strategy.address, 5_000, {"from": gov})
-    chain.sleep(1)
-    strategy.harvest({"from": strategist})
-    half = int(amount / 2)
-    assert pytest.approx(strategy.estimatedTotalAssets(), rel=RELATIVE_APPROX) == half
+    tokenB.approve(router, 2 ** 256 - 1, {"from": tokenB_whale})
+    router.swapExactTokensForTokens(
+        tokenB.balanceOf(tokenB_whale),
+        0,
+        [tokenB, tokenA],
+        tokenB_whale,
+        2 ** 256 - 1,
+        {"from": tokenB_whale},
+    )
 
+    print(
+        f"Joint estimated assets: {joint.estimatedTotalAssetsInToken(tokenA) / 1e18} {tokenA.symbol()} and {joint.estimatedTotalAssetsInToken(tokenB) / 1e18} {tokenB.symbol()}"
+    )
 
-def test_sweep(gov, vault, strategy, token, user, amount, weth, weth_amount):
-    # Strategy want token doesn't work
-    token.transfer(strategy, amount, {"from": user})
-    assert token.address == strategy.want()
-    assert token.balanceOf(strategy) > 0
-    with brownie.reverts("!want"):
-        strategy.sweep(token, {"from": gov})
+    # Wait plz
+    chain.sleep(3600 * 24)
+    chain.mine(int(3600 * 24 / 13) - 30)
 
-    # Vault share token doesn't work
-    with brownie.reverts("!shares"):
-        strategy.sweep(vault.address, {"from": gov})
+    print(
+        f"Joint estimated assets: {joint.estimatedTotalAssetsInToken(tokenA) / 1e18} {tokenA.symbol()} and {joint.estimatedTotalAssetsInToken(tokenB) / 1e18} {tokenB.symbol()}"
+    )
 
-    # TODO: If you add protected tokens to the strategy.
-    # Protected token doesn't work
-    # with brownie.reverts("!protected"):
-    #     strategy.sweep(strategy.protectedToken(), {"from": gov})
+    # If there is any profit it should go to the providers
+    assert joint.pendingReward() > 0
+    # If joint doesn't reinvest, and providers do not invest want, the want
+    # will stay in the providers
+    providerA.setInvestWant(False, {"from": strategist})
+    providerB.setInvestWant(False, {"from": strategist})
+    providerA.setTakeProfit(True, {"from": strategist})
+    providerB.setTakeProfit(True, {"from": strategist})
+    providerA.harvest({"from": strategist})
+    providerB.harvest({"from": strategist})
 
-    before_balance = weth.balanceOf(gov)
-    weth.transfer(strategy, weth_amount, {"from": user})
-    assert weth.address != strategy.want()
-    assert weth.balanceOf(user) == 0
-    strategy.sweep(weth, {"from": gov})
-    assert weth.balanceOf(gov) == weth_amount + before_balance
+    assert providerA.balanceOfWant() > 0
+    assert providerB.balanceOfWant() > 0
 
+    lossA = vaultA.strategies(providerA).dict()["totalLoss"]
+    lossB = vaultB.strategies(providerB).dict()["totalLoss"]
 
-def test_triggers(chain, gov, vault, strategy, token, amount, user, strategist):
-    # Deposit to the vault and harvest
-    actions.user_deposit(user, vault, token, amount)
-    vault.updateStrategyDebtRatio(strategy.address, 5_000, {"from": gov})
-    chain.sleep(1)
-    strategy.harvest()
+    assert lossA > 0
+    assert lossB > 0
 
-    strategy.harvestTrigger(0)
-    strategy.tendTrigger(0)
+    returnA = -lossA / investedA
+    returnB = -lossB / investedB
+
+    print(
+        f"Return: {returnA*100:.5f}% {tokenA.symbol()} {returnB*100:.5f}% {tokenB.symbol()}"
+    )
+
+    assert pytest.approx(returnA, rel=50e-3) == returnB
